refactor: split insights API module by responsibility
